usb_device, usb_desc: updating it to be MISRA compliant
- usb_device, usb_desc - updating it to be MISRA compliant

+++ Bootloader/STM32/Src/usb_device.c
@@ -76,7 +76,7 @@ MX_USB_DEVICE_Init(void) {
     if (USBD_RegisterClass(&hUsbDeviceFS, &USBD_CDC) != USBD_OK) {
         Error_Handler();
     }
-    if (USBD_CDC_RegisterInterface(&hUsbDeviceFS, &USBD_Interface_fops_FS) != USBD_OK) {
+    if (USBD_CDC_RegisterInterface(&hUsbDeviceFS, &USBD_Interface_fops_FS) != (uint8_t)USBD_OK) {
         Error_Handler();
     }
     if (USBD_Start(&hUsbDeviceFS) != USBD_OK) {
